Extract AndroidProject class into a separate .js file
