+++ cordova-lib/src/plugman/platforms/android.js
@@ -23,7 +23,7 @@ var fs = require('fs')  // use existsSync in 0.6.x
    , events = require('../events')
    , xml_helpers = require(path.join(__dirname, '..', '..', 'util', 'xml-helpers'))
    , properties_parser = require('properties-parser')
-   , shell = require('shelljs');
+   , android_project = require('../util/android-project');
 
 var projectFileCache = {};
 
@@ -115,84 +115,13 @@ module.exports = {
     },
     parseProjectFile: function(project_dir){
         if (!projectFileCache[project_dir]) {
-            projectFileCache[project_dir] = {
-                propertiesEditors: {},
-                subProjectDirs : {},
-                addSubProject: function(parentDir, subDir) {
-                    var subProjectFile = path.resolve(subDir, "project.properties");
-                    if (!fs.existsSync(subProjectFile)) throw new Error('cannot find "' + subProjectFile + '" referenced in <framework>');
-
-                    var parentProjectFile = path.resolve(parentDir, "project.properties");
-                    var parentProperties = this._getPropertiesFile(parentProjectFile);
-                    addLibraryReference(parentProperties, module.exports.getRelativeLibraryPath(parentDir, subDir));
-
-                    var subProperties = this._getPropertiesFile(subProjectFile);
-                    subProperties.set("target", parentProperties.get("target"));
-
-                    this.subProjectDirs[subDir] = true;
-                    this._dirty = true;
-                },
-                removeSubProject: function(parentDir, subDir) {
-                    var parentProjectFile = path.resolve(parentDir, "project.properties");
-                    var parentProperties = this._getPropertiesFile(parentProjectFile);
-                    removeLibraryReference(parentProperties, module.exports.getRelativeLibraryPath(parentDir, subDir));
-                    delete this.subProjectDirs[subDir];
-                    this._dirty = true;
-                },
-                write: function () {
-                    if (!this._dirty) return;
-
-                    for (var filename in this.propertiesEditors) {
-                        fs.writeFileSync(filename, this.propertiesEditors[filename].toString());
-                    }
-
-                    for (var sub_dir in this.subProjectDirs)
-                    {
-                        shell.exec("android update lib-project --path " + sub_dir);
-                    }
-                    this._dirty = false;
-                },
-                _dirty : false,
-                _getPropertiesFile: function (filename) {
-                    if (!this.propertiesEditors[filename])
-                        this.propertiesEditors[filename] = properties_parser.createEditor(filename);
-
-                    return this.propertiesEditors[filename];
-                }
-            };
+            projectFileCache[project_dir] = new android_project.AndroidProject();
         }
 
         return projectFileCache[project_dir];
     },
     purgeProjectFileCache:function(project_dir) {
         delete projectFileCache[project_dir];
-    },
-    getRelativeLibraryPath: function (parentDir, subDir) {
-        var libraryPath = path.relative(parentDir, subDir);
-        return (path.sep == '\\') ? libraryPath.replace(/\\/g, '/') : libraryPath;
     }
 };
 
-function addLibraryReference(projectProperties, libraryPath) {
-    var i = 1;
-    while (projectProperties.get("android.library.reference." + i))
-        i++;
-
-    projectProperties.set("android.library.reference." + i, libraryPath);
-}
-
-function removeLibraryReference(projectProperties, libraryPath) {
-    var i = 1;
-    var currentLib;
-    while (currentLib = projectProperties.get("android.library.reference." + i)) {
-        if (currentLib === libraryPath) {
-            while (currentLib = projectProperties.get("android.library.reference." + (i + 1))) {
-                projectProperties.set("android.library.reference." + i, currentLib);
-                i++;
-            }
-            projectProperties.set("android.library.reference." + i);
-            break;
-        }
-        i++;
-    }
-}
